+++ cpp/src/io/orc/reader_impl.cu
@@ -602,9 +602,10 @@ reader::impl::impl(std::unique_ptr<datasource> source,
   _decimals_as_int_scale = options.forced_decimals_scale;
 }
 
-std::unique_ptr<table> reader::impl::read(int skip_rows, int num_rows,
-                                          int stripe, cudaStream_t stream) {
+table_with_metadata reader::impl::read(int skip_rows, int num_rows, int stripe,
+                                       cudaStream_t stream) {
   std::vector<std::unique_ptr<column>> out_columns;
+  table_metadata out_metadata;
 
   // Select only stripes required (aka row groups)
   const auto selected_stripes =
@@ -768,7 +769,17 @@ std::unique_ptr<table> reader::impl::read(int skip_rows, int num_rows,
     }
   }
 
-  return std::make_unique<table>(std::move(out_columns));
+  // Return column names (must match order of returned columns)
+  out_metadata.column_names.resize(_selected_columns.size());
+  for (size_t i = 0; i < _selected_columns.size(); i++) {
+    out_metadata.column_names[i] = _metadata->ff.GetColumnName(_selected_columns[i]);
+  }
+  // Return user metadata
+  for (const auto& kv : _metadata->ff.metadata) {
+    out_metadata.user_data.insert({kv.name, kv.value});
+  }
+
+  return { std::make_unique<table>(std::move(out_columns)), std::move(out_metadata) };
 }
 
 // Forward to implementation
@@ -793,20 +804,20 @@ reader::reader(std::shared_ptr<arrow::io::RandomAccessFile> file,
 reader::~reader() = default;
 
 // Forward to implementation
-std::unique_ptr<table> reader::read_all(cudaStream_t stream) {
+table_with_metadata reader::read_all(cudaStream_t stream) {
   return _impl->read(0, -1, -1, stream);
 }
 
 // Forward to implementation
-std::unique_ptr<table> reader::read_stripe(size_type stripe,
-                                           cudaStream_t stream) {
+table_with_metadata reader::read_stripe(size_type stripe,
+                                        cudaStream_t stream) {
   return _impl->read(0, -1, stripe, stream);
 }
 
 // Forward to implementation
-std::unique_ptr<table> reader::read_rows(size_type skip_rows,
-                                         size_type num_rows,
-                                         cudaStream_t stream) {
+table_with_metadata reader::read_rows(size_type skip_rows,
+                                      size_type num_rows,
+                                      cudaStream_t stream) {
   return _impl->read(skip_rows, (num_rows != 0) ? num_rows : -1, -1, stream);
 }